Fixing warnings revealed by xcode (clang).
Copied from Perforce
 Change: 178898
 ServerID: perforce.ravenbrook.com

+++ mps/code/mv2test.c
@@ -76,7 +76,7 @@ static float ran1(long *idum)
   *idum=IA*(*idum-k*IQ)-IR*k;   /* Compute idum=(IA*idum) % IM without
                                    overflows by Schrage's method. */
   if (*idum < 0) *idum += IM;
-  j=iy/NDIV;                    /* Will be in the range 0..NTAB-1. */
+  j=(int)(iy/NDIV);             /* Will be in the range 0..NTAB-1. */
   iy=iv[j];                     /* Output previously stored value and
                                    refill the shuffle table. */
   iv[j] = *idum;
